+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 
 ### Breaking changes
 
-* Adding animation, vertical_align, word_break, display, visibility, & position arguments to the utilities class.
+* Adding animation, vertical_align, word_break, display, visibility, & position arguments to the utilities class. `animation: :grow` is now `animation: :hover_grow` this was a change because we changed the class name in primer.
 
     *Jon Rohan*
+
+### Misc
+
+* Prevent flexible tag syntax with rubocop rule.
+
+    *Kate Higa*
 
 ## 0.0.45