+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### Misc
 
+* Update contributing guidelines with release instructions.
+
+    *Kate Higa*
+
 * Prevent flexible tag syntax with rubocop rule.
 
     *Kate Higa*